chore: move realtime router module
